Update feature freeze information.
There will be no Codo 2 based on CoffeeScript Redux, at least not for now.

To be honest, I'm a bit confused about Redux. I always had the impression
that Redux will be the official CoffeeScript V2, but it turns out this is not
the case for sure. It might happen sometimes, but not anytime soon. So the
current situation is that we now have two different CoffeeScript versions
that are incompatible in certain areas.

Since I'm anyways very short on time and need to reduce my time spent on
my various OSS projects, I officially cancel my plan for Codo 2 with this commit
and I'm not going to implement more features anytime soon.

* Closes #16
* Closes #72
* Closes #79
* Closes #81
* Closes #84
* Closes #89
* Closes #91
* Closes #94

If you'd like to see any of those features in Codo, please open a pull request.