Avoid reversing text order in @property comment
Currently the text following the type declaration in the `@property`
statement was appended to the rest of the comment. This caused the
resulting comment to be very confusing if there was any other text in
the comment, as the contents of the first line were appended to the very
end.