Fix empty property description hiding documentation
PR #229 introduced an issue where if the @property's description wasn't
filled in, the rest of the comment body would also be left out of the
documentation. This was caused by `property[2]` being undefined and
prepended to the `lines` array which caused the while loop to terminate
early. The while loop checks if the last value shifted from the array is
`undefined` not if there are any more values left in the array.